Tidy and consolidate pomodoro code
